Edit solution
By Dejan Svetoslavov
