+++ Seminars/Sem.03/Pract.03/Task02.cpp
@@ -1,177 +1,148 @@
 #include <iostream>
 #include <fstream>
 
-using std::cin, std::cout, std::endl, std::ofstream, std::ifstream;
+const int MAX_LENGHT = 25;
 
-const char *FILE_NAME = "jobOffers.dat";
-const int max_length = 25;
-
-struct jobOffer {
-    char companyName[max_length + 1];
-    size_t teamSize;
-    size_t vacationDays;
-    long long salary;
+struct JobOffer {
+    char companyName[MAX_LENGHT + 1];
+    size_t numberOfDevelopers;
+    size_t numberOfVacationDays;
+    long long priceOfProject;
 };
 
-void printJobOffer(const jobOffer &job) {
-    cout << "Company name: " << job.companyName << endl;
-    cout << "Team size: " << job.teamSize << endl;
-    cout << "Vacation days: " << job.vacationDays << endl;
-    cout << "Salary: " << job.salary << endl;
-    cout << "-----------------------------------" << endl;
+size_t Mystrcmp(const char* str1, const char* str2) {
+    while (*str1 && (*str1 == *str2)) {
+        ++str1;
+        ++str2;
+    }
+    return *(const unsigned char*) str1 - *(const unsigned char*) str2;
 }
 
-bool writeJobOffersToFile(const char *&filename, const jobOffer *const &jobs, const int &N) {
-    ofstream file(filename, std::ios::binary | std::ios::app);
-    if (!file.is_open()) {
-        cout << "File could not be opened!" << endl;
-        return false;
-    }
-    for (int i = 0; i < N; ++i) {
-        file << jobs[i].companyName << " " << jobs[i].teamSize << " " << jobs[i].vacationDays << " " << jobs[i].salary
-             << endl;
-    }
+void ReadJobOffer(JobOffer& jobOffer) {
+    std::cout << "Enter details for job offer:" << std::endl;
 
-    cout << "Job offers successfully written to file!" << endl;
+    std::cout << "Company name: ";
+    std::cin.getline(jobOffer.companyName, MAX_LENGHT);
 
-    file.close();
-    return true;
+    std::cout << "Number of developers: ";
+    std::cin >> jobOffer.numberOfDevelopers;
+
+    std::cout << "Number of vacation days: ";
+    std::cin >> jobOffer.numberOfVacationDays;
+
+    std::cout << "Price of project: ";
+    std::cin >> jobOffer.priceOfProject;
+
+    std::cin.ignore(); // ignore the newline character
 }
 
-
-void readJobsFromTerminal(bool multiple = false) {
-    int N = 1;
-    if (multiple) {
-        cout << "Enter number of jobs: ";
-        cin >> N;
-    }
-    jobOffer *jobs = new jobOffer[N];
-    for (int i = 0; i < N; ++i) {
-        cout << "Enter company name: ";
-        cin >> jobs[i].companyName;
-        cout << "Enter team size: ";
-        cin >> jobs[i].teamSize;
-        cout << "Enter vacation days: ";
-        cin >> jobs[i].vacationDays;
-        cout << "Enter salary: ";
-        cin >> jobs[i].salary;
-        cout << endl;
-    }
-
-    writeJobOffersToFile(FILE_NAME, jobs, N);
-    delete[] jobs;
+void PrintJobOffer(const JobOffer& jobOffer) {
+    std::cout << "Company Name: " << jobOffer.companyName << std::endl;
+    std::cout << "Number of Developers: " << jobOffer.numberOfDevelopers << std::endl;
+    std::cout << "Number of Vacation Days: " << jobOffer.numberOfVacationDays << std::endl;
+    std::cout << "Price of Project: " << jobOffer.priceOfProject << std::endl;
 }
 
-void filterOffers(const char *filename, long long minSalary) {
-    ifstream file(filename, std::ios::binary);
-    if (!file.is_open()) {
-        cout << "File could not be opened!" << endl;
-        return;
+void StoreJobOffer(const JobOffer& jobOffer, const char* fileName) {
+    std::ofstream file(fileName, std::ios::binary | std::ios::app);
+    if (file.is_open()) {
+        file.write((const char*) &jobOffer, sizeof(JobOffer));
+        file.close();
+    } else {
+        std::cout << "Unable to open file for writing." << std::endl;
     }
-    jobOffer job{};
-    while (file >> job.companyName >> job.teamSize >> job.vacationDays >> job.salary) {
-        if (job.salary >= minSalary) {
-            printJobOffer(job);
+}
+
+void listAllJobOffers(const char* fileName) {
+    std::ifstream file(fileName, std::ios::in | std::ios::binary);
+    if (file.is_open()) {
+        JobOffer jobOffer;
+        while (file.read((char*) &jobOffer, sizeof(JobOffer))) {
+                PrintJobOffer(jobOffer);
+                std::cout << std::endl;
         }
+        file.close();
+    } else {
+        std::cout << "Unable to open file for reading." << std::endl;
     }
-
-    file.close();
 }
 
-bool existsOffer(const char *filename, const char *companyName) {
-    ifstream file(filename, std::ios::binary);
-    if (!file.is_open()) {
-        cout << "File could not be opened!" << endl;
-        return false;
-    }
-    jobOffer job{};
-    int counter = 0;
-    while (file >> job.companyName >> job.teamSize >> job.vacationDays >> job.salary) {
-        if (strcmp(job.companyName, companyName) == 0) {
-            counter++;
-            printJobOffer(job);
+void ListJobOffersWithGreaterPriceThan(long long projectPrice, const char* fileName) {
+    std::ifstream file(fileName, std::ios::binary);
+    if (file.is_open()) {
+        JobOffer jobOffer;
+        while (file.read((char*) &jobOffer, sizeof(JobOffer))) {
+            if (jobOffer.priceOfProject > projectPrice) {
+                PrintJobOffer(jobOffer);
+                std::cout << std::endl;
+            }
         }
+        file.close();
+    } else {
+        std::cout << "Unable to open file for reading." << std::endl;
     }
-
-    file.close();
-    return counter;
 }
 
-void printAllOffers(const char *filename) {
-    ifstream file(filename, std::ios::binary);
-    if (!file.is_open()) {
-        cout << "File could not be opened!" << endl;
-        return;
-    }
-    jobOffer job{};
-    while (file >> job.companyName >> job.teamSize >> job.vacationDays >> job.salary) {
-        printJobOffer(job);
-    }
-
-    file.close();
-}
-
-void perfectOffer(const char *filename, int maxCoworkers, int minVacancyDays, int minSalary) {
-    ifstream file(filename, std::ios::binary);
-    if (!file.is_open()) {
-        cout << "File could not be opened!" << endl;
-        return;
-    }
-    jobOffer job{};
-    int counter = 0;
-    while (file >> job.companyName >> job.teamSize >> job.vacationDays >> job.salary) {
-        if (job.teamSize <= maxCoworkers && job.vacationDays >= minVacancyDays && job.salary >= minSalary) {
-            counter++;
-            printJobOffer(job);
+void ListJobOffersFromCompany(const char* companyName, const char* fileName) {
+    std::ifstream file(fileName,std::ios::binary);
+    if (file.is_open()) {
+        JobOffer jobOffer;
+        while (file.read((char*) &jobOffer, sizeof(JobOffer))) {
+            if (Mystrcmp(jobOffer.companyName, companyName) == 0) {
+                PrintJobOffer(jobOffer);
+                std::cout << std::endl;
+            }
         }
+        file.close();
+    } else {
+        std::cout << "Unable to open file for reading." << std::endl;
     }
+}
 
-    if (counter == 0) {
-        cout << "No matching offers!" << endl;
-    }
-
-    file.close();
+void DisplayMenu() {
+    std::cout << "Job Offers Menu:" << std::endl;
+    std::cout << "1. Input and store in a binary file a job offer" << std::endl;
+    std::cout << "2. List all job offers" << std::endl;
+    std::cout << "3. List all job offers that have priceOfProject greater than n" << std::endl;
+    std::cout << "4. List all job offers from a company" << std::endl;
+    std::cout << "5. Exit" << std::endl;
 }
 
 int main() {
-    cout << "Welcome to the job offers database!" << endl;
-    cout << "Supported commands:" << endl;
-    cout << "   a - add job offer" << endl;
-    cout << "   i - print all job offers" << endl;
-    cout << "   s <company_name> - search for job offer" << endl;
-    cout << "   f <salary> - filter job offers" << endl;
-    cout << "   q - quit" << endl;
-    char c = ' ';
-    while (c != 'q') {
-        cout << endl << "Enter command: ";
-        cin >> c;
-        cout << endl;
-        switch (c) {
-            case 'a':
-                readJobsFromTerminal();
+    int choice = -1;
+    JobOffer currentJobOffer;
+    char companyName[MAX_LENGHT + 1];
+    long long minProjectPrice = 0;
+    while (choice != 0) {
+        DisplayMenu();
+        std::cout << "Enter your choice: ";
+        std::cin >> choice;
+        std::cin.ignore();
+        switch (choice) {
+            case 1:
+                std::cout << "Option 1 selected" << std::endl;
+                ReadJobOffer(currentJobOffer);
+                StoreJobOffer(currentJobOffer, "jobOffers.bin");
                 break;
-            case 'i':
-                printAllOffers(FILE_NAME);
+            case 2:
+                std::cout << "Option 2 selected" << std::endl;
+                ListAllJobOffers("jobOffers.bin");
                 break;
-            case 's':
-                cin.ignore(1, ' ');
-                char companyName[max_length + 1];
-                cin.getline(companyName, max_length + 1);
-                if (!existsOffer(FILE_NAME, companyName)) {
-                    cout << "Company doesn't exists!" << endl;
-                }
+            case 3:
+                std::cout << "Company name: ";
+                std::cin >> minProjectPrice;
+                ListJobOffersWithGreaterPriceThan(minProjectPrice, "jobOffers.bin");
                 break;
-            case 'f':
-                cin.ignore(1, ' ');
-                long long minSalary;
-                cin >> minSalary;
-                filterOffers(FILE_NAME, minSalary);
+            case 4:
+                std::cout << "Company name: ";
+                std::cin.getline(companyName, 25);
+                ListJobOffersFromCompany(companyName, "jobOffers.bin");
                 break;
-            case 'q':
-                cout << "Exiting..." << endl;
+            case 0:
+                std::cout << "Exiting..." << std::endl;
                 break;
             default:
-                cout << "Invalid command!" << endl;
+                std::cout << "Invalid choice. Please try again." << std::endl;
                 break;
         }
     }
